requirements.txt
pyautogui
fastapi
pillow
jinja2
uvicorn[standard]